Fix logger, more comments, updated readme
