--- gosyncmodules/CheckForError.go
@@ -1,8 +1,0 @@
-package gosyncmodules
-
-func CheckForError(e error) {
-	if e != nil {
-		logger.Errorln(e)
-		panic(e)
-	}
-}
