+++ syncer/ADData.go
@@ -1,7 +1,8 @@
-package gosyncmodules
+package syncer
 
 type keyvalue map[string]interface{}
 
+// LDAPElement holds the DN and attributes of an LDAP/AD entry.
 type LDAPElement struct {
 	DN         string
 	attributes []keyvalue
