+++ syncer/ConvertRealmToLower.go
@@ -1,4 +1,4 @@
-package gosyncmodules
+package syncer
 
 import (
 	"gopkg.in/ldap.v2"
@@ -6,15 +6,18 @@ import (
 	"strings"
 )
 
-//WrapperStruct to embed *ldap.AddRequest and add a custom method to it.
+// AddRequest is a wrapperStruct to embed *ldap.AddRequest and add a custom method to it.
 type AddRequest struct {
 	*ldap.AddRequest
 }
 
+// SetDN is used to set the Dn in an AddRequest
 func (a *AddRequest) SetDN(dn string) {
 	a.DN = dn
 }
 
+// ConvertRealmToLower is a normalisation function for Windows Active Directory. This is required because
+// the realm is returned capitalised in Active Directory, and needs to be normalised for the sake of LDAP.
 func ConvertRealmToLower(upperrealm []*ldap.AddRequest) {
 
 	r := regexp.MustCompile(`[A-Z]+=`)
@@ -29,6 +32,7 @@ func ConvertRealmToLower(upperrealm []*ldap.AddRequest) {
 
 }
 
+// ConvertAttributesToLower is a normalisation function. This is for the sake of LDAP
 func ConvertAttributesToLower(upperAttribute *[]string) *[]string {
 	r := regexp.MustCompile(`[A-Z]+=`)
 	var attributeAggregated []string
