+++ syncer/FindUniques.go
@@ -1,12 +1,14 @@
-package gosyncmodules
+package syncer
 
 import (
 	//	"github.com/nohupped/ldap"
 	"gopkg.in/ldap.v2"
 )
 
+// Action map is a lame attempt to define if the action needs to be an Add or a Delete
 type Action map[string]*ldap.AddRequest
 
+// FindAdds function is used to find the entries that needs to be modified/added to the slave
 func FindAdds(ADElementsConverted, LDAPElementsConverted *[]*ldap.AddRequest, LdapConnection *ldap.Conn, AddChan chan Action, shutdownAddChan chan string) {
 	logger.Debugln("Starting FindAdds")
 	defer func() { shutdownAddChan <- "Done from func FindAdds" }()
@@ -19,13 +21,13 @@ func FindAdds(ADElementsConverted, LDAPElementsConverted *[]*ldap.AddRequest, Ld
 			CheckAttributes(LdapConnection, LDAPEntry, i)
 			continue
 		} else {
-			//err := LDAPConnection.Add(i)
 			AddChan <- Action{"Add": i} //Write composite literal to channel
 
 		}
 	}
 }
 
+// FindDels function is used to find the entries that needs to be deleted from the slave
 func FindDels(LDAPElementsConverted, ADElementsConverted *[]*ldap.AddRequest, DelChan chan Action, shutdownDelChan chan string) {
 	logger.Debugln("Starting FindDels")
 	defer func() { shutdownDelChan <- "Done from func FindDels" }()
@@ -37,7 +39,6 @@ func FindDels(LDAPElementsConverted, ADElementsConverted *[]*ldap.AddRequest, De
 			continue
 		} else {
 			logger.Debugln(i.DN, "Doesn't exist in AD, will be set to delete.")
-			//err := LDAPConnection.Add(i)
 			DelChan <- Action{"Del": i} //Write composite literal to channel
 
 		}
