+++ syncer/sampleConfig.go
@@ -1,5 +1,6 @@
-package gosyncmodules
+package syncer
 
+// SampleConfig is used to print a sample config to STDOUT with all the required fields.
 const SampleConfig = `### Sample config generated by the program. Edit accordingly ###
 [ADServer]
 Host = <host ip>
